+++ lib/track.dart
@@ -3,4 +3,6 @@ library;
 export 'core/base_counter_tracker.dart';
 export 'core/base_tracker.dart';
 export 'core/tracker_period.dart';
+export 'extensions/adapter_list_extensions.dart';
 export 'services/history_tracker.dart';
+export 'services/activity_counter.dart';